poller: heartbeat in the module the CLI actually runs (codex review of #87)
bin/cli.mjs imports src/team-relay/room-poller.mjs; the heartbeat had
landed in src/room-poller.mjs, an unimported legacy copy, so production
'rooms watch' would never have touched it and the nudge gate would have
suppressed every wake after 180 s. Moved; the legacy file is back to
untouched. New test starts the real startRoomPoller against a stub curl
on PATH and checks the heartbeat is written by the first poll and
refreshed by the interval poll.
