+++ test/poller-health.test.mjs
@@ -2,7 +2,7 @@
 // Issue #86: a nudge must not fire while the room poller is down, and the
 // supervisor must say so once (and once more when it is back).
 import assert from 'node:assert/strict';
-import { mkdtempSync, rmSync, writeFileSync, existsSync, utimesSync, readFileSync } from 'node:fs';
+import { mkdtempSync, rmSync, writeFileSync, existsSync, utimesSync, readFileSync, chmodSync } from 'node:fs';
 import { tmpdir } from 'node:os';
 import path from 'node:path';
 import { spawnSync, execFile } from 'node:child_process';
@@ -11,7 +11,7 @@ const execFileP = promisify(execFile);
 import { createServer } from 'node:http';
 import test from 'node:test';
 import { fileURLToPath } from 'node:url';
-import { writeHeartbeat } from '../src/room-poller.mjs';
+import { writeHeartbeat, startRoomPoller } from '../src/team-relay/room-poller.mjs';
 
 const repoRoot = path.resolve(path.dirname(fileURLToPath(import.meta.url)), '..');
 const nudge = path.join(repoRoot, 'tools', 'codex_gui_nudge.sh');
@@ -84,4 +84,32 @@ test('poller-health-alert posts once when down, once when back, nothing in betwe
 
     await run(); assert.equal(posts.length, 2, 'healthy + no state = silent');
   } finally { server.close(); rmSync(dir, { recursive: true, force: true }); }
+});
+
+test('the poller the CLI actually runs writes the heartbeat on every poll', async () => {
+  // src/team-relay/room-poller.mjs is what bin/cli.mjs imports; a stub curl on
+  // PATH keeps it off the network (same trick as the ioreg stub tests).
+  const dir = mkdtempSync(path.join(tmpdir(), 'iak-live-'));
+  const stub = path.join(dir, 'curl');
+  writeFileSync(stub, '#!/bin/sh\necho "[]"\n');
+  chmodSync(stub, 0o755);
+  const savedPath = process.env.PATH;
+  process.env.PATH = `${dir}:${savedPath}`;
+  const hb = path.join(dir, 'hb');
+  let timers;
+  try {
+    timers = await startRoomPoller({
+      rooms: ['r'], apiKey: 'k', handle: '@t', interval: 1,
+      config: { poller: { seen_file: path.join(dir, 'seen'), notification_file: path.join(dir, 'notify'), heartbeat_file: hb, nudge_mode: 'none' }, queue: { path: path.join(dir, 'q.jsonl') } }
+    });
+    assert.ok(existsSync(hb), 'heartbeat written by the first poll');
+    const first = readFileSync(hb, 'utf8');
+    await new Promise((r) => setTimeout(r, 1300));
+    assert.notEqual(readFileSync(hb, 'utf8'), first, 'heartbeat refreshed by the interval poll');
+  } finally {
+    if (timers?.roomTimer) clearInterval(timers.roomTimer);
+    if (timers?.dmTimer) clearInterval(timers.dmTimer);
+    process.env.PATH = savedPath;
+    rmSync(dir, { recursive: true, force: true });
+  }
 });
